# OpenAPI.yaml
openapi: 3.0.1
info:
  title: Weather
  description: Проект по получению данных о погоде по названию города
  contact:
    name: Antonov Vladimir
    url: https://t.me/bloodyt3ars
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: weather
    description: В этом разделе можно получить сведение о погоде в желаемом городе
paths:
  /weather/{cityName}:
    get:
      tags:
        - weather
      description: Получить погоду по названию города
      operationId: getWeather
      parameters:
        - name: cityName
          in: path
          description: Название города
          required: true
          schema:
            type: string
          example: Санкт-Петербург
        - name: serviceName
          in: query
          description: Название сервиса погоды
          required: false
          schema:
            type: string
          example: yandexweather
      responses:
        '200':
          description: Все прошло успешно, город найден, погода получена
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Некорректный запрос
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherDto'
components:
  schemas:
    WeatherDto:
      type: object
      properties:
        city:
          type: string
          description: Название города
          example: Санкт-Петербург
        latitude:
          type: string
          description: Широта
          example: '59.938732'
        longitude:
          type: string
          description: Широта
          example: '30.316229'
        temperature:
          type: string
          description: Температура
          example: '21.04'
        description:
          type: string
          description: Описание
          example: Небольшая облачность